Download crawl def in own function
Simpler to use the existing get_report for job and only have a specialized function for crawl definition, rather than mixing crawl job and crawl definition into the same function.
